+++ test/elastix_test.exs
@@ -1,3 +1,0 @@
-defmodule ElastixTest do
-  use ExUnit.Case
-end