+++ test/rubber_test.exs
@@ -1,0 +1,3 @@
+defmodule RubberTest do
+  use ExUnit.Case
+end